Complete task TA612 of user story US363.
Complete the implementation and testing of adding the resource manager's
features by using class MemBuffer as the mechanism to test. The test case
is temporary and currently resides within group GrpBasicInit inside of test
DisableCompletely.

The integrity of allocations was verified using the valgrind app.

Documentation was created to outline how to add a new group to the test
framework.

Various documentation improvements were made.