Add Singlton class CtrlrConfig.
The global Singleton design pattern for class CtrlrConfig is complete and
unit tested. The unit test is temporary. The class is the sole interface to
the Controller Configuration (CC) register. This class will aid the setup
after a power up situation configure the controller before ASQ and ACQ's are
even created.