Reimplement meta data buffers more efficiently.
The lifetime of a meta data buffer was thought of in terms of a cmd's lifetime.
However it was noticed that extra calls into the kernel we being done to
release that back to the system. And if new cmds needed meta data buffers then
more calls to allocate memory which was just perhaps released would be done.

It was realized that the "releasing" of meta data buffers could be delayed until
the end of group lifetime was complete. This avoids freeing and allocating if
we simply hold on to the buffer in user space, and then issue it back out for
new objects desiring meta data buffers. Essentially we have placed a memory
allocation layer in user space to cut down on the calls into kernel space.