Add issue of identify cmd to admin SQ.
In group GrpBasicInit test case Identify cmd has been coded to issue the
controller data structure and the namespace structure. The cmds are issued
to the ASQ and reaped via the ACQ. The data returned is dumped to a file.

The test framework is mostly complete at this point. The only remaining items
are to create the rest of the commands for all command sets.