Bugfix: ignore errors could cause kernel core dump.
It was realized that ignoring of errors was supported, but that it could cause
a kernel core dump if the error was bad enough. This didn't mean the feature
couldn't be supported, rather if dnvme and tnvme could re-sync are do a soft
reset after the error then the core dump could be elleviated and we would still
ignore errors. Thus, ignoring of errors no means, that the next test within
the next group can run if application. It does NOT mean that the next test
within the same group can run, this could lead to assumptions and cause a core
dump withint the framework.

Tests within groups are logically related, subsequent test rely on priort ones
passing, so if an error occurs, subsequent test will make an illegal assumption.