Major rearchitecture; Addresses number of queues.
The number of queues a DUT supports is not provided by a registger. Rather it
must be set, and when set it remains that value until a power cycle according
to specification. Factor in the problems encountered when various tests
need that value set differently we conclude this needs to be settable from
the command line.

It was discovered that all tests need some fundamental knowledge about the DUT.
Things like what is the SQ size, how many Q's are supported, how many
namespaces are supported. Then factor in that each test or group of tests
will be requesting this data, it makes sense to gather this and provide it
to all tests since this is non-volatile data. Thus the introduction of the
gInformative singleton was coded.