Bugfix: Test didn't correctly consider discontig mem.
The hdw may not support discontig mem to back an IOQ. GrpBasicInit is written
generically to not fail if the hdw doesn't support discontig mem. However this
wasn't coded correctly, although the IOSQ/IOCQ's wouldn't have been created,
subsequent tests were not checking properly and started to write/read to queues
which cannot be created. All tests now check for discontig support before
running the discontig part of the test.