Addressed the logging system producing large files.
If tnvme was left unchecked it would have produced massive amounts of dump files
in the subdirectories created for each group. Inquiry into a number of
validation teams indicated that data assoc with failed tests is mainly of
importance. Hence, now <root_log_dir>/GrpInformative contains the data
extracted from the DUT upon app instantiation. While <root_log_dir>/GrpPending
contains the resource dump files for the last executed group only; each new
group destroys the dump files from the prior group.

An additional efficiency was introduced when the stdout/stderr of tnvme was
redirected to a logging utility. tnvme speed improvements of 8 fold are being
experienced. The logging utility, svlogd, needs to be installed to take
advantage of this feature. It rotates the stdout/stderr from tnvme into
a set of output files. If and when a test fails you will want to inspect file
<root_log_dir>/current for the details, this file always contains the last
output from tnvme before it is rotated. For an example of this improvement
reference file manage/runtnvme.sh from the manage repo under the github
compliancesuite project.