Auto-retrieve get log page error log when a CE fails.
The framework allows helper methods to detect if a given CE is reported a
specified (SC, SCT) value pair. When the M bit is set this indicates more
information is available in the error log page. This data could help in
deciphering how a test failed, whether the hdw is misbehaving or the test case
is incorrectly coded. After a test fails all the data to help in the
investigation should be available. After all this is the job of the framework,
to automate those things which every test will desire.

The framework will now dump the contents of the error log when a test case
fails. The framework doesn't care if the test failed due to a CE indicating
failure and the M bit is set, rather the log page is retrieved anyway just in
case is will be useful.