Avoid PRP misuse, enhance the framework to implement constraints.
A given PRP buffer for any cmd may be set such the dnvme obeys the manufacturing
of a user space buffer into PRP lists and pointers. A test developer could set
an illegal case, by accident or on purpose. But the spec states there is nothing
to gain by this, since the outcome is unknown. The framework will, during an
individual cmd instantation, set how the PRP pointer are allowed to be
interpreted according to the spec. Later when the cmd is used and a test
developer states how a buffer should be interpreted is must be within the
allowed range of settings applied during instantiation.
See method tnvme/Cmds/prpData.h/SetPrpBuffer() and method SetPrpAllowed().
As a further example see tnvme/Cmds/createIOCQ.cpp/CreateIOCQ() constructor.