Remove queues element size assumptions.
The IOSQ and IOCQ classes defined static variables to state the size of a
common element. Since the "common" or required element size is defined as a
constant in the NVMe spec. It makes far more sense to use the constant as
reported by the device for maximum test coverage. This solution is also
scalable/dynamic and tests the device's implementation of the spec.