Bugfix: Segmentation fault during test failure.
The possibility of a tnvme segmentation fault after a test fails is possible,
because tests use to "throw exception()"; this could break out of
local test lifetime scope. These resources would be freed, but hdw may be
unaware that it is in violation of some test criteria and still be active and
could continue to write to memory backing an IOCQ for example, which could
have been freed.

For this reason we can no longer except test failures by returning boolean
values, now the framework forces a special object to be thrown. This object
called Exception/FrmwkEx handles disabling the DUT and dumping important
debug within the local scope to keep the resources alive until such time the
DUT is disable/deactivated.