stdout wasn't being flushed when needed.
The LOG_XXX() macros direct to stderr which is flush immediately. Some
statements, early in tnvme's launch must be printf's. The buffering of stdout
wasn't be merged correctly with the logging statements making it very difficult
to debug issues. Couldn't see a reason the app couln't launch, for example,
when you forget to change the permissions bits on the /dev/nvme0 special file.