Changed the naming of 'logging' to 'dumping'.
The name change is to prevent confusion. The so called logging output directory
was not the logging directory at all. Rather is was the place where all tnvme's
dump files will reside. The logging is controlled by macro's like LOG_NRM() and
LOG_DBG() and these are sent to stderr. Thus in the pure sense 'logging' is
thus stderr and it was likely people would become confused. They are really
2 separate concepts.