Small bug fix.
tnvme didn't properly report the number of queues supported by a device, it
suffered from a rollover of a 16 bit variable. Then added a warning to flag
an inherent problem in the NVMe spec. Namely....

This warning is to make people aware that although the Set Features cmd
with ID=0x07 is based upon a 0-based number and theoretically states a
device could support 0x10000 IOQ's, that in fact a device can never
create that many. The Create IOQ cmd's only allow 16 bits to
represent a QID and those are 1-based values. Thus QID 0x10000 is not
possible and therefore 0x10000 IOQ's are not possible since IOQ's
are defined to be those from 1 to 0xffff.