Removed safeguard checking.
The removal was due to rethinking the linkage between cmds and the gInformative
singleton. Accessing a singlton isn't necessary because asking for more DWORDs
is, yes, undefined, but the value of the data is undefined, not the outcome. So
in other words the kernel won't crash and one may even want to request more
data that is available for testing purposes.