Remove the singleton requirement, "no throwing".
No throwing from a singleton was necessary for keep the main() function from
having to wrap all code with a try catch block. In hind sight this is an
unreasonable limitation to place upon the source. Throwing is powerfull and
it a cornerstop of main stream c++. It is being allowed now.