Redesigned RsrcMngr class.
The interaction between dnvme and tnvme is a tighly coupled marriage. Anyone
out of sync with the other causes problems. The fact the gRcrcMngr allowed 3
flavors of freeing resources caused some people to start using them, which is
what a public interface is for, right? Correct, but in this case the methods
come with strings attached, and someone without a deep understanding of the
source layering could cause memory leaks in the kernel. This programming hole
has been filled.