Bugfix: CID in cmd was not updated.
When sending a cmd into the driver, dnvme, it will assigned a CID, unique
cmd ID when a cmd is copied into an SQ. The dnvme then passes knownledge
of that CID back to tnvme. The CID was never set to the cmd for later
retrieval, i.e. the bug.