Updated Cmd class to reuse code instead of copy-paste.  Also added set/get nibble functions.
Updated identify command for better support of CNS values.

Fixed namespace attach command to have correct PRP bitmask.

Added utility codes to GrpAdminNamespaceManagement.

Added bit definitions for the VS register.

Added BasicAdminQueueSetup function to Queues which has the code common to may tests for the initial setup of the admin queues.  This function can be used in place of that code for less copy-paste.